examples/webcode2m/requirements.txt
tqdm
requests==2.27
pillow
bs4
tinycss2
playwright
pyquery
cssutils
datasets
transformers
torch
wandb
aiohttp
aiofiles
urllib3==1.26.11
"git+https://github.com/openai/CLIP"
scikit-learn
scikit-image